setup.sh: fix the logic of running prctl externally for umu
somehow, i was completely blind to the fact that calling prctl from umu with the helper would do nothing once that exited, since it only applies to children of the process making the prctl call.

fix this by making the call from the wrapper and then running umu-run from that.